refactor: Centralize weekly analysis logic in WeeklyAnalysisService
